break out ResolveNonEmpty
