+++ main/core/src/mill/define/Module.scala
@@ -118,8 +118,8 @@ object Module {
           .getClasses
           .filter(implicitly[ClassTag[T]].runtimeClass.isAssignableFrom(_))
           .flatMap { c =>
-            c.getSimpleName match{
-              case s"$name$$" if filter(name)=>
+            c.getSimpleName match {
+              case s"$name$$" if filter(name) =>
                 c.getFields.find(_.getName == "MODULE$").map(_.get(c).asInstanceOf[T])
               case _ => None
             }
