+++ main/core/src/mill/eval/Evaluator.scala
@@ -822,8 +822,8 @@ object Evaluator {
               val Segment.Label(tName) = segments.value.last
               Segments(
                 segments.value.init ++
-                Seq(Segment.Label(tName + ".super")) ++
-                t.ctx.enclosing.split("[.# ]").map(Segment.Label)
+                  Seq(Segment.Label(tName + ".super")) ++
+                  t.ctx.enclosing.split("[.# ]").map(Segment.Label)
               )
             }
           )
