+++ main/src/mill/main/RunScript.scala
@@ -10,95 +10,16 @@ import mill.api.Strict.Agg
 import scala.reflect.ClassTag
 import mill.define.ParseArgs.TargetsWithParams
 
-/**
- * Custom version of ammonite.main.Scripts, letting us run the build.sc script
- * directly without going through Ammonite's main-method/argument-parsing
- * subsystem
- */
 object RunScript {
 
   type TaskName = String
 
-
-  def resolveTasks[T, R: ClassTag](
-      resolver: mill.main.Resolver[R],
-      evaluator: Evaluator,
-      scriptArgs: Seq[String],
-      selectMode: SelectMode
-  ): Either[String, List[R]] = {
-    val parsedGroups: Either[String, Seq[TargetsWithParams]] = ParseArgs(scriptArgs, selectMode)
-    val resolvedGroups = parsedGroups.flatMap { groups =>
-      val resolved = groups.map { parsed: TargetsWithParams =>
-        resolveTasks(resolver, evaluator, Right(parsed))
-      }
-      EitherOps.sequence(resolved)
-    }
-    resolvedGroups.map(_.flatten.toList)
-  }
-
-  private def resolveTasks[T, R: ClassTag](
-      resolver: mill.main.Resolver[R],
-      evaluator: Evaluator,
-      targetsWithParams: Either[String, TargetsWithParams]
-  ): Either[String, List[R]] = {
-    for {
-      parsed <- targetsWithParams
-      (selectors, args) = parsed
-      taskss <- {
-        val selected = selectors.map { case (scopedSel, sel) =>
-          for (rootModule <- resolveRootModule(evaluator, scopedSel))
-            yield {
-
-              try {
-                // We inject the `evaluator.rootModule` into the TargetScopt, rather
-                // than the `rootModule`, because even if you are running an external
-                // module we still want you to be able to resolve targets from your
-                // main build. Resolving targets from external builds as CLI arguments
-                // is not currently supported
-                mill.eval.Evaluator.currentEvaluator.set(evaluator)
-                resolver.resolve(
-                  sel.value.toList,
-                  rootModule,
-                  rootModule.millDiscover,
-                  args
-                )
-              } finally {
-                mill.eval.Evaluator.currentEvaluator.set(null)
-              }
-            }
-        }
-        EitherOps.sequence(selected).map(_.toList)
-      }
-      res <- EitherOps.sequence(taskss)
-    } yield res.flatten
-  }
-
-  def resolveRootModule[T](evaluator: Evaluator, scopedSel: Option[Segments]) = {
-    scopedSel match {
-      case None => Right(evaluator.rootModule)
-      case Some(scoping) =>
-        for {
-          moduleCls <-
-            try Right(evaluator.rootModule.getClass.getClassLoader.loadClass(scoping.render + "$"))
-            catch {
-              case e: ClassNotFoundException =>
-                Left("Cannot resolve external module " + scoping.render)
-            }
-          rootModule <- moduleCls.getField("MODULE$").get(moduleCls) match {
-            case rootModule: ExternalModule => Right(rootModule)
-            case _ => Left("Class " + scoping.render + " is not an external module")
-          }
-        } yield rootModule
-    }
-  }
-
-
-  def evaluateTasks[T](
+  def evaluateTasks(
       evaluator: Evaluator,
       scriptArgs: Seq[String],
       selectMode: SelectMode
   ): Either[String, (Seq[PathRef], Either[String, Seq[(Any, Option[ujson.Value])]])] = {
-    for (targets <- resolveTasks(ResolveTasks, evaluator, scriptArgs, selectMode))
+    for (targets <- ResolveTasks.resolveTasks(evaluator, scriptArgs, selectMode))
       yield {
         val (watched, res) = evaluate(evaluator, Agg.from(targets.distinct))
 
@@ -117,7 +38,7 @@ object RunScript {
       scriptArgs: Seq[String],
       selectMode: SelectMode
   ): Either[String, (Seq[PathRef], Either[String, Seq[(Any, Option[(TaskName, ujson.Value)])]])] = {
-    for (targets <- resolveTasks(ResolveTasks, evaluator, scriptArgs, selectMode))
+    for (targets <- ResolveTasks.resolveTasks(evaluator, scriptArgs, selectMode))
       yield {
         val (watched, res) = evaluateNamed(evaluator, Agg.from(targets.distinct))
